Replaces best_hyperparameters
The best_hyperparameters.py script is limited because it doesn't understand max/min logic. That is, it cannot report which run achieved the highest validation accuracy, only which had the highest final validation accuracy. However, you can retrieve this information from the W&B website.

Therefore we modify this script to take a run as argument and return its hyperparameters. Sample use:

     ./get_hyperparameters.py cuny-cl/ewt/cvyrtb6d

This returns the hyperparameters for that run, which I previously determined to be the best using W&B's run spreadsheet.
