+++ examples/wandb_sweeps/README.md
@@ -57,3 +57,8 @@ Then, one can retrieve the results as follows:
 
 3.  Click on the downward arrow link, select "CSV Export", then click "Save as
     CSV".
+
+Or, to get the hyperparameters for a particular run, copy the "Run path" from
+the run's "Overview" on W&B, and then run:
+
+    ./get_hyperparameters.py "${RUN_PATH}"
